Move htmx error to views
